extract buffer from sender
